ecom-django/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.2
django-ckeditor==6.4.0
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
idna==3.4
phonenumbers==8.13.31
pillow==10.2.0
python-dotenv==1.0.1
requests==2.28.2
sqlparse==0.4.3
stripe==5.0.0
urllib3==1.26.14
